Change to EasyBoard.cs
Change to the CrewCanBoard function. Was originally using this function for my mod testing and found a way that was a lot more certain as to if a Kerbal can board.

I was getting NulRefExcept in my code because your original CrewCanBoard function was saying it was OK to board when it wasn't. Hope this little change I found helps. :)

Adam